requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==4.4.2
filelock==3.14.0
fsspec==2024.5.0
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
importlib-metadata==7.1.0
jinja2==3.1.4
llvmlite==0.41.1
MarkupSafe==2.1.5
more-itertools==10.2.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
pillow==10.3.0
proglog==0.1.10
pytube==15.0.0
regex==2024.5.15
requests==2.32.2
sympy==1.12
tiktoken==0.7.0
torch==2.2.2
tqdm==4.66.4
typing-extensions==4.12.0
urllib3==2.2.1
zipp==3.19.0
